Add files via upload
An active control technique termed “flossing” that uses a dual nanopore device is presented to trap a proteintagged DNA molecule and up to 100's of back‐and‐forth electrical scans of the molecule are performed in a few seconds. The protein motifs bound to 48.5 kb λ‐DNA are used as detectable features for active triggering of the bidirectional control. Molecular noise is suppressed by averaging the multiscan data to produce averaged intertag distance estimates that are comparable to their known values. Since nanopore feature‐mapping applications require DNA linearization when passing through the pore, a key advantage of flossing is that trans‐pore linearization is increased to >98% by the second scan, compared to 35% for single nanopore passage of the same set of molecules. In concert with barcoding methods, the dual‐pore flossing technique could enable genome mapping and structural variation applications, or mapping loci of epigenetic relevance.