Add files via upload
An insulator-based dielectrophoresis (iDEP) is a label-free method that has been extensively utilized
for manipulation of nanoparticles, cells, and biomolecules. Here, we present a new iDEP approach that
can rapidly trap nanoparticles at the close proximity of a glass nanopipette’s tip by applying 10V/cm
direct current (DC) across the pipette’s length. The trapping mechanism was systemically studied using
both numerical modeling and experimental observations. The results showed that the particle trapping
was determined to be controlled by three dominant electrokinetic forces including dielectrophoretic,
electrophoretic and electroosmotic force. Furthermore, the efect of the ionic strength, the pipette’s
geometry, and the applied electric feld on the entrapment efciency was investigated. To show the
application of our device in biomedical sciences, we demonstrated the successful entrapment of
fuorescently tagged liposomes and unlabeled plasma-driven exosomes from the PBS solution. Also, to
illustrate the selective entrapment capability of our device, 100nm liposomes were extracted from the
PBS solution containing 500nm polystyrene particles at the tip of the pipette as the voltage polarity
was reversed.